[tz::core] tz::handle now has the underlying type uint32_t, meaning it can be used directly in structs exposed to shaders, equivalent to a tzsl `uint`

+++ src/tz/core/data/handle.hpp
@@ -6,7 +6,7 @@
 
 namespace tz
 {
-	enum class hanval : std::size_t{};
+	enum class hanval : std::uint32_t{};
 
 	struct nullhand_t{};
 	constexpr nullhand_t nullhand;
